requirements.txt
asgiref==3.4.0
certifi==2021.5.30
charset-normalizer==2.0.3
Django==3.2.4
django-appconf==1.0.4
django-ckeditor==6.1.0
django-compressor==2.4.1
django-js-asset==1.2.2
django-libsass==0.8
django-modeltranslation==0.17.3
django-rosetta==0.7.6
django-social-share==2.2.1
django-tinymce==3.3.0
flake8==3.9.2
idna==3.2
libsass==0.21.0
mccabe==0.6.1
microsofttranslator==0.5
Pillow==8.3.0
polib==1.1.1
pycodestyle==2.7.0
pydocstyle==6.1.1
pyflakes==2.3.1
python-decouple==3.4
pytz==2021.1
rcssmin==1.0.6
requests==2.26.0
rjsmin==1.1.0
six==1.16.0
snowballstemmer==2.1.0
sqlparse==0.4.1
urllib3==1.26.6
